enables Java-like usage pattern
The simplest use case is not supported anymore – use Enumeration for
that. Now, it is always required to subclass Enum.Val (rather than just
aliasing the type).

It is also allowed now to define the Value subclass
outside the Enum object, usually as the companion class of the Enum
object, which feels much closer to Java Enum because the inner class
Enum.Val does not get exposed, anymore (e.g. in method signatures).